# authman-samba-contract/src/main/swagger/samba_connector.yml
swagger: "2.0"
info:
  title: "Samba 4 Connector"
  version: "1.0.0"
  description: "Samba 4 connector interface."
  contact:
    name: "Christian Bremer"
    email: "bremersee@googlemail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "base-api.bremersee.org"
basePath: "/smb-con"
...
schemes:
  - "http"
  - "https"
paths:
  /api/v1/info:
    get:
      tags:
        - "samba-connector-controller"
      summary: "Get samba server info."
      description: "Get samba server info."
      operationId: "getInfo"
      produces:
        - "application/json"
      responses:
        200:
          description: "The samba server info."
          schema:
            $ref: "#/definitions/Info"
      security:
        - bremersee_auth:
          - "samba:admin"
  /api/v1/groups:
    get:
      tags:
        - "samba-connector-controller"
      summary: "Get all samba groups."
      description: "Get all samba groups."
      operationId: "getGroups"
      produces:
        - "application/json"
      responses:
        200:
          description: "A list with samba group items."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/SambaGroupItem"
      security:
        - bremersee_auth:
          - "samba:admin"
    post:
      tags:
        - "samba-connector-controller"
      summary: "Add a samba group."
      operationId: "addGroup"
      parameters:
        - name: "group"
          description: "The samba group to add."
          in: "body"
          schema:
            $ref: "#/definitions/SambaGroup"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "The added samba group."
          schema:
            $ref: "#/definitions/SambaGroup"
        400:
          description: "Bad request"
      security:
        - bremersee_auth:
          - "samba:admin"
  /api/v1/groups/{groupName}:
    get:
      tags:
      - "samba-connector-controller"
      summary: "Get a samba group by name."
      operationId: "getGroupByName"
      parameters:
        - name: "groupName"
          description: "The samba group name."
          type: "string"
          required: true
          in: "path"
      produces:
        - "application/json"
      responses:
        200:
          description: "The samba group with the specified name."
          schema:
            $ref: "#/definitions/SambaGroup"
        404:
          description: "Not found"
      security:
        - bremersee_auth:
          - "samba:admin"
    delete:
      tags:
        - "samba-connector-controller"
      summary: "Delete a samba group."
      operationId: "deleteGroup"
      parameters:
        - name: "groupName"
          description: "The samba group name."
          type: "string"
          required: true
          in: "path"
      produces:
        - "application/json"
      responses:
        400:
          description: "Bad request"
      security:
        - bremersee_auth:
          - "samba:admin"
  /api/v1/groups/{groupName}/members:
    put:
      tags:
        - "samba-connector-controller"
      summary: "Update samba group members."
      operationId: "updateGroupMembers"
      parameters:
        - name: "groupName"
          description: "The samba group name."
          type: "string"
          required: true
          in: "path"
        - name: "members"
          description: "The samba group members."
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Names"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "The updated samba group."
          schema:
            $ref: "#/definitions/SambaGroup"
        400:
          description: "Bad request"
      security:
        - bremersee_auth:
          - "samba:admin"
  /api/v1/users:
    post:
      tags:
        - "samba-connector-controller"
      summary: "Add a samba user."
      operationId: "addUser"
      parameters:
        - name: "sambaUser"
          description: "The samba user to add."
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/SambaUserAddRequest"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "The added samba user."
          schema:
            $ref: "#/definitions/SambaUser"
        400:
          description: "Bad request"
      security:
        - bremersee_auth:
          - "samba:admin"
  /api/v1/users/{userName}:
    get:
      tags:
        - "samba-connector-controller"
      summary: "Get a samba user by name."
      operationId: "getUser"
      parameters:
        - name: "userName"
          description: "The user name of the samba user."
          type: "string"
          required: true
          in: "path"
      produces:
        - "application/json"
      responses:
        200:
          description: "The samba user with the specified name."
          schema:
            $ref: "#/definitions/SambaUser"
        404:
          description: "User not found"
      security:
        - bremersee_auth:
          - "samba:admin"
    patch:
      tags:
        - "samba-connector-controller"
      summary: "Updates a samba user."
      operationId: "updateUser"
      parameters:
        - name: "userName"
          description: "The user name of the samba user."
          type: "string"
          required: true
          in: "path"
        - name: "sambaUser"
          description: "The update data of the samba user."
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/SambaUser"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "The updated samba user."
          schema:
            $ref: "#/definitions/SambaUser"
        400:
          description: "Bad request"
        404:
          description: "User not found"
      security:
        - bremersee_auth:
          - "samba:admin"
    delete:
      tags:
        - "samba-connector-controller"
      summary: "Deletes a samba user."
      operationId: "deleteUser"
      parameters:
        - name: "userName"
          description: "The user name of the samba user."
          type: "string"
          required: true
          in: "path"
      produces:
        - "application/json"
      responses:
        400:
          description: "Bad request"
        404:
          description: "User not found"
      security:
        - bremersee_auth:
          - "samba:admin"
  /api/v1/users/{userName}/f/exists:
    get:
      tags:
        - "samba-connector-controller"
      summary: "Checks whether a samba user exists."
      operationId: "userExists"
      parameters:
        - name: "userName"
          description: "The user name of the samba user."
          type: "string"
          required: true
          in: "path"
      produces:
        - "application/json"
      responses:
        200:
          description: "True if the user exists otherwise false."
          schema:
            $ref: "#/definitions/BooleanWrapper"
      security:
        - bremersee_auth:
          - "samba:admin"
  /api/v1/users/{userName}/groups:
    put:
      tags:
        - "samba-connector-controller"
      summary: "Updates samba user groups."
      operationId: "updateUserGroups"
      parameters:
        - name: "userName"
          description: "The user name of the samba user."
          type: "string"
          required: true
          in: "path"
        - name: "groups"
          description: "The groups of the samba user."
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Names"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "The updated samba user."
          schema:
            $ref: "#/definitions/SambaUser"
        400:
          description: "Bad request"
        404:
          description: "User not found"
      security:
        - bremersee_auth:
          - "samba:admin"
  /api/v1/users/{userName}/password:
    put:
      tags:
        - "samba-connector-controller"
      summary: "Updates the password of the samba user."
      operationId: "updateUserPassword"
      parameters:
        - name: "userName"
          description: "The user name of the samba user."
          type: "string"
          required: true
          in: "path"
        - name: "newPassword"
          description: "The password of the samba user."
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Password"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "The password was successfully changed."
        400:
          description: "Bad request"
        404:
          description: "User not found"
      security:
        - bremersee_auth:
          - "samba:admin"
  /api/v1/dns/zones:
    get:
      tags:
        - "samba-connector-controller"
      summary: "Get all nameserver zones."
      description: "Get all nameserver zones."
      operationId: "getDnsZones"
      produces:
        - "application/json"
      responses:
        200:
          description: "A list with nameserver zones."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/DnsZone"
      security:
        - bremersee_auth:
          - "samba:admin"
    post:
      tags:
        - "samba-connector-controller"
      summary: "Create new nameserver zone."
      description: "Create new nameserver zone."
      operationId: "createDnsZone"
      parameters:
        - name: "request"
          in: "body"
          schema:
            $ref: "#/definitions/DnsZoneCreateRequest"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "The zone was successfully created."
      security:
        - bremersee_auth:
          - "samba:admin"
    delete:
      tags:
        - "samba-connector-controller"
      summary: "Delete nameserver zone."
      description: "Delete nameserver zone."
      operationId: "deleteDnsZone"
      parameters:
        - name: "zoneName"
          in: "query"
          type: "string"
          required: true
      produces:
        - "application/json"
      responses:
        200:
          description: "The zone was successfully created."
      security:
        - bremersee_auth:
          - "samba:admin"
  /api/v1/dns/zones/records:
    get:
      tags:
        - "samba-connector-controller"
      summary: "Get all nameserver records of a zone."
      description: "Get all nameserver records of a zone."
      operationId: "getDnsRecords"
      parameters:
        - name: "zoneName"
          description: "The nameserver zone name."
          type: "string"
          required: true
          in: "query"
      produces:
        - "application/json"
      responses:
        200:
          description: "A list with nameserver records."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/DnsEntry"
      security:
        - bremersee_auth:
          - "samba:admin"
    post:
      tags:
        - "samba-connector-controller"
      summary: "Create or delete nameserver record."
      description: "Create or delete nameserver record."
      operationId: "createOrDeleteDnsRecord"
      parameters:
        - name: "action"
          in: "query"
          type: string
          enum:
            - "CREATE"
            - "DELETE"
          required: true
        - name: "request"
          in: "body"
          schema:
            $ref: "#/definitions/DnsRecordRequest"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "The nameserver record was successfully created."
      security:
        - bremersee_auth:
          - "samba:admin"
    put:
      tags:
        - "samba-connector-controller"
      summary: "Update nameserver record."
      description: "Update nameserver record."
      operationId: "updateDnsRecord"
      parameters:
        - name: "request"
          in: "body"
          schema:
            $ref: "#/definitions/DnsRecordUpdateRequest"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "The nameserver record was successfully created."
      security:
        - bremersee_auth:
          - "samba:admin"
securityDefinitions:
  bremersee_auth:
    type: "oauth2"
    authorizationUrl: "http://localhost:8080/authman/oauth/authorize"
    tokenUrl: "http://localhost:8080/authman/oauth/token"
    flow: "password"
    scopes:
      samba:admin: "Administer the Samba AD."
definitions:
  BooleanWrapper:
    description: "Wrapper of a boolean value."
    type: "object"
    properties:
      value:
        type: "boolean"
        description: "The boolean value."
    required:
      - value
  Info:
    description: "A description of the samba domain controller."
    type: "object"
    properties:
      nameServerHost:
        type: "string"
        description: "The host name of the name server."
  Name:
    description: "A simple name or a distinguished name."
    type: "object"
    properties:
      value:
        type: "string"
        description: "The name."
      distinguishedName:
        type: "boolean"
        default: "false"
        description: "Is the name a distinguished name?"
    required:
      - value
  Names:
    type: "object"
    description: "A list with names."
    properties:
      values:
        type: "array"
        items:
          $ref: "#/definitions/Name"
        description: "The names."
  Common:
    type: "object"
    description: "Common attributes"
    properties:
      distinguishedName:
        type: "string"
        description: "The distinguished name in the active directory."
      created:
        type: "string"
        format: "date-time"
        description: "The creation date."
      modified:
        type: "string"
        format: "date-time"
        description: "The last modification date."
    discriminator: "_type"
  SambaGroupItem:
    description: "Samba group item."
    allOf:
      - $ref: "#/definitions/Common"
      - type: "object"
        properties:
          name:
            type: "string"
            description: "The name of the samba group."
    discriminator: "_type"
  SambaGroup:
    description: "Samba group."
    allOf:
      - $ref: "#/definitions/SambaGroupItem"
      - type: "object"
        properties:
          members:
            type: "array"
            items:
              $ref: "#/definitions/Name"
            description: "The members of the samba group."
  SambaUser:
    description: "Samba user."
    allOf:
      - $ref: "#/definitions/Common"
      - type: "object"
        properties:
          userName:
            type: "string"
            description: "The user name of the samba user."
          enabled:
            type: "boolean"
            description: "Is the samba user enabled?"
          passwordLastSet:
            type: "string"
            format: "date-time"
          displayName:
            type: "string"
          email:
            type: "string"
            format: "email"
            description: "The mail address of the samba user."
          mobile:
            type: "string"
            description: "The mobile number of the samba user."
          groups:
            type: "array"
            items:
              $ref: "#/definitions/Name"
            description: "The groups of the samba user."
  SambaUserAddRequest:
    description: "Samba user to add."
    allOf:
      - $ref: "#/definitions/SambaUser"
      - type: "object"
        properties:
          password:
            type: "string"
            format: "password"
            description: "The password of the samba user."
  Password:
    type: "object"
    properties:
      value:
        type: "string"
        format: "password"
        description: "A password."
    required:
      - value
  DnsZoneFlag:
    type: "object"
    description: "DNS Zone Flag"
    properties:
      name:
        type: "string"
        description: "The zone flag name."
    required:
      - name
  DnsDwDpZoneFlag:
    type: "object"
    description: "DNS Zone Flag"
    properties:
      name:
        type: "string"
        description: "The zone flag name."
    required:
      - name
  DnsZone:
    type: "object"
    description: "DNS Zone"
    properties:
      pszZoneName:
        type: "string"
        description: "The zone name."
      flags:
        type: "array"
        items:
          $ref: "#/definitions/DnsZoneFlag"
        description: "The zone flags."
      zoneType:
        type: "string"
        description: "The zone type."
      version:
        type: "string"
        description: "The version."
      dwDpFlags:
        type: "array"
        items:
          $ref: "#/definitions/DnsDwDpZoneFlag"
        description: "The dwDp zone flags."
      pszDpFqdn:
        type: "string"
        description: "The pszDpFqdn."
  DnsZoneCreateRequest:
    type: "object"
    description: "DNS Zone Create Request"
    properties:
      pszZoneName:
        type: "string"
        description: "The zone name."
    required:
      - pszZoneName
  DnsRecordType:
    type: "string"
    description: "DNS Record Types."
    enum:
      - "A"
      - "AAAA"
      - "PTR"
      - "CNAME"
      - "NS"
      - "MX"
      - "SRV"
      - "TXT"
      - "SOA"
      - "AFSDB"
      - "APL"
      - "CAA"
      - "CDNSKEY"
      - "CDS"
      - "CERT"
      - "DHCID"
      - "DLV"
      - "DNAME"
      - "DNSKEY"
      - "DS"
      - "HIP"
      - "IPSECKEY"
      - "KEY"
      - "KX"
      - "LOC"
      - "NAPTR"
      - "NSEC"
      - "NSEC3"
      - "NSEC3PARAM"
      - "OPENPGPKEY"
      - "RRSIG"
      - "RP"
      - "SIG"
      - "SSHFP"
      - "TA"
      - "TKEY"
      - "TLSA"
      - "TSIG"
      - "URI"
  DnsRecordRequest:
    type: "object"
    description: "DNS Record Create Request"
    properties:
      zoneName:
        type: "string"
        description: "The zone name."
      name:
        type: "string"
        description: "The record name."
      recordType:
        $ref: '#/definitions/DnsRecordType'
      data:
        type: "string"
        description: "The record data."
    required:
      - zoneName
      - name
      - recordType
      - data
  DnsRecordUpdateRequest:
    type: "object"
    description: "DNS Record Create Request"
    properties:
      zoneName:
        type: "string"
        description: "The zone name."
      name:
        type: "string"
        description: "The record name."
      recordType:
        $ref: '#/definitions/DnsRecordType'
      oldData:
        type: "string"
        description: "The old record data."
      newData:
        type: "string"
        description: "The new record data."
    required:
      - zoneName
      - name
      - recordType
      - oldData
      - newData
  DnsRecord:
    type: "object"
    description: "DNS Record"
    properties:
      recordType:
        type: "string"
        description: "The nameserver record type."
      recordValue:
        type: "string"
        description: "The nameserver record value."
      flags:
        type: "string"
        description: "The nameserver record flags."
      serial:
        type: "string"
        description: "The nameserver record serial."
      ttl:
        type: "string"
        description: "The nameserver record ttl."
  DnsEntry:
    type: "object"
    description: "DNS Record"
    properties:
      name:
        type: "string"
        description: "The entry name."
      records:
        type: "array"
        items:
          $ref: "#/definitions/DnsRecord"
        description: "The nameserver records."